# api/doc/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/add-user": {
      "post": {
        "operationId": "UserController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The user has been successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/login": {
      "post": {
        "operationId": "UserController_login",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/get-user/{id}": {
      "get": {
        "operationId": "UserController_getUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/logout": {
      "post": {
        "operationId": "UserController_logout",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ]
      }
    }
  },
  "info": {
    "title": "FCRM",
    "description": "The list of APIs",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "The username of the user",
            "example": "john_doe",
            "minLength": 3
          },
          "email": {
            "type": "string",
            "description": "The email address of the user",
            "example": "john.doe@example.com"
          },
          "password": {
            "type": "string",
            "description": "The password for the user account",
            "example": "securePassword123",
            "minLength": 6
          }
        },
        "required": [
          "username",
          "email",
          "password"
        ]
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique identifier of the user",
            "example": "c9b1e3d0-3b5d-4f5e-8e16-d5a9ebecb9b6"
          },
          "username": {
            "type": "string",
            "description": "The username of the user",
            "example": "john_doe"
          },
          "email": {
            "type": "string",
            "description": "The email address of the user",
            "example": "john.doe@example.com"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "The date when the user was created",
            "example": "2024-09-08T10:43:04.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "The date when the user was last updated",
            "example": "2024-09-08T10:43:04.000Z"
          }
        },
        "required": [
          "id",
          "username",
          "email",
          "createdAt",
          "updatedAt"
        ]
      }
    }
  }
}